modal delete

+++ src/app/inventories/components/provider/provider-service/provider-service.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject, OnInit, TemplateRef, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { MainContainerComponent } from 'ngx-dabd-grupo01';
+import { ConfirmAlertComponent, MainContainerComponent } from 'ngx-dabd-grupo01';
 import { ProviderServiceUpdateComponent } from "../provider-service-update/provider-service-update.component";
 import { ServiceService } from '../../../services/suppliers/service.service';
 import { Service } from '../../../models/suppliers/service.model';
@@ -109,14 +109,23 @@ export class ProviderServiceComponent implements OnInit {
   }
 
   deleteService(id: number) {
-    this.serviceService.deleteService(id).subscribe({
-      next: () => {
-        this.loadServices();
-      },
-      error: (error) => {
-        console.error('Error deleting service:', error);
+    const modalresult = this.modalService.open(ConfirmAlertComponent); 
+    modalresult.componentInstance.alertTitle = 'Confirmación';
+    modalresult.componentInstance.alertMessage = '¿Está seguro que desea eliminar este servicio?';
+    modalresult.componentInstance.alertVariant = 'delete';
+
+    modalresult.result.then((result) => {
+      if(result){
+        this.serviceService.deleteService(id).subscribe({
+          next: () => {
+            this.loadServices();
+          },
+          error: (error) => {
+            console.error('Error deleting service:', error);
+          }
+        });
       }
-    });
+    })
   }
 
     openModalFilter(): void {
